requirements.txt
feedparser>=6.0.10
requests>=2.31.0
trafilatura>=1.12.0
python-dotenv>=1.0.0
google-genai>=1.0.0
supabase>=2.9.0
resend>=2.5.0
pydub>=0.25.1
static-ffmpeg>=3.0
audioop-lts>=0.2.2
